requirements.txt
Flask==3.0.2
requests==2.31.0
python-dotenv==1.0.1
Werkzeug==3.0.1
Jinja2==3.1.3
MarkupSafe==2.1.5
itsdangerous==2.1.2
click==8.1.7
google-generativeai==0.8.3
moviepy==1.0.3
Pillow==10.2.0
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
opencv-python==4.8.1.78
gunicorn==21.2.0